NOJIRA: Several cleanups and small bugs in the scripts MH-3534 : The 'curl' issue is solved and the problems with our card are partially fixed MH-3280 : We were able to do a full round trip, from scheduling to ingestion, with success MH-3553 : This is currently handled now, the java-6-sun is set up as default
git-svn-id: https://opencast.jira.com/svn/MH/trunk@6816 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ docs/scripts/ubuntu_capture_agent/setup_source.sh
@@ -10,30 +10,43 @@ if [[ ! $INSTALL_RUN ]]; then
     exit 1
 fi
 
-# Get the necessary matterhorn source code (the whole trunk, as specified in MH-3211)
-unset answer
-read -p "Do you wish to download the source code from the official trunk? (Y/n): " answer
-while [[ -z "$(echo "${answer:-Y}" | grep -i '^[yn]')" ]]; do
-    read -p "Please enter a valid answer (Y/n): " answer
-done
-
-if [[ $(echo "${answer:-Y}" | grep -i '^[n]') ]]; then
-    read -p "Enter the complete download address (default: trunk): " address
+# Detect if the matterhorn source has been already checked out
+url=$(svn info $TRUNK 2> /dev/null | grep URL: | cut -d ' ' -f 2) 
+if [[ -n "$url" ]]; then
+    read -p "The source $url has been already checked out. Do you wish to keep it (Y/n)? " keep
+    while [[ -z "$(echo "${keep:-Y}" | grep -i '^[yn]')" ]]; do
+	read -p "Please enter a valid answer (Y/n): " keep
+    done  
+else
+    keep=no
 fi
 
-echo -n "Downloading matterhorn source from repository... "
-rm -rf $TRUNK
-svn co --force ${address:-$TRUNK_URL} $TRUNK > /dev/null
-if [[ $? -ne 0 ]]; then
-    echo "Error!"
-    echo "Couldn't check out the matterhorn code. Aborting"
-    exit 1
+if [[ -n "$(echo "${keep:-Y}" | grep -i '^n')" ]]; then
+    # Get the necessary matterhorn source code (the whole trunk, as specified in MH-3211)
+    unset answer
+    read -p "Do you wish to download the source code from the official trunk? (Y/n): " answer
+    while [[ -z "$(echo "${answer:-Y}" | grep -i '^[yn]')" ]]; do
+	read -p "Please enter a valid answer (Y/n): " answer
+    done
+    
+    if [[ $(echo "${answer:-Y}" | grep -i '^[n]') ]]; then
+	read -p "Enter the complete download address (default: trunk): " address
+    fi
+    
+    echo -n "Downloading matterhorn source from repository... "
+    rm -rf $TRUNK
+    svn co --force ${address:-$TRUNK_URL} $TRUNK > /dev/null
+    if [[ $? -ne 0 ]]; then
+	echo "Error!"
+	echo "Couldn't check out the matterhorn code. Aborting"
+	exit 1
+    fi
+    echo "Done"
 fi
-echo "Done"
-
 
 # Setup felix configuration
 echo -n "Applying matterhorn configuration files to felix... "
-cp -r $TRUNK/docs/felix/bin/* ${FELIX_HOME}/bin
-cp -r $TRUNK/docs/felix/conf ${FELIX_HOME}
+
+cp -rf $TRUNK/docs/felix/bin ${FELIX_HOME}
+cp -rf $TRUNK/docs/felix/conf ${FELIX_HOME}
 echo "Done"
